Simplified GUI for image_correct_json_generate
A simplified GUI for generating image correction configuration file. It has the same function of "image_correct_json_generate.py", but it assumes files of the same group are exclusively in the same directory. It does not provide all the options, and it is rather an example of how the configuration is generated.